+++ src/components/Form/DropdownSelect/index.tsx
@@ -1,1 +1,0 @@
-export { DropdownSelect, type DropdownSelectProps } from "./DropdownSelect";