+++ src/components/Form/Select/constants.ts
@@ -1,55 +1,26 @@
-import { ColourPropTypes, CommonAndHTMLProps } from "@/components/Element/constants";
-import React from "react";
+import { CommonAndHTMLProps } from "@/components/Element/constants";
 
 export type SelectElementType = HTMLSelectElement;
-export type SelectWithSearchElementType = HTMLSelectElement;
 
 export type OptionElementType = HTMLOptionElement;
 
 export interface OptionProps {
-    value: string;
-    label: string;
-    disabled?: boolean;
-    selected?: boolean;
-}
-
-export interface OptionForSearchWithSelectProps {
-    value: string;
-    label: string;
-    customLabel?: React.ReactNode;
-    disabled?: boolean;
+    value      : string;
+    label      : string;
+    disabled ? : boolean;
+    selected ? : boolean;
 }
 
 export interface OptGroupProps {
-    label: string;
-    options: OptionProps[];
+    label   : string;
+    options : OptionProps[];
 }
 
 export interface SelectCustomProps {
-    options: OptionProps[];
-    label?: string;
-    helpText?: string;
-    errorText?: string;
-}
-
-export interface SelectWithSearchCustomProps {
-    options: OptionForSearchWithSelectProps[];
-    label?: string;
-    helpText?: string;
-    errorText?: string;
-    allowMultiSelect?: boolean;
-    onChange?: (value: string | string[]) => void;
-    disabled?: boolean;
-    placeholder?: string;
-    id?: string;
-    defaultValue?: string;
-    badgeBgColour?: ColourPropTypes;
-    badgeBgColor?: ColourPropTypes;
-    badgeTextColour?: ColourPropTypes;
-    badgeTextColor?: ColourPropTypes;
-    selectionLimit?: number;
-    allowCustomEntries?: boolean;
+    options     : OptionProps[];
+    label     ? : string;
+    helpText  ? : string;
+    errorText ? : string;
 }
 
 export type SelectProps = Omit<CommonAndHTMLProps<SelectElementType>, keyof SelectCustomProps> & SelectCustomProps;
-export type SelectWithSearchProps = Omit<CommonAndHTMLProps<SelectWithSearchElementType>, keyof SelectWithSearchCustomProps> & SelectWithSearchCustomProps;