+++ src/components/index.tsx
@@ -60,8 +60,6 @@ export {
     type InputLabelProps,
     Select,
     type SelectProps,
-    SelectWithSearch,
-    type SelectWithSearchProps,
     FileUpload,
     type FileUploadProps,
     RadioGroup,
@@ -78,8 +76,8 @@ export {
     type RadioTabGroupProps,
     PinInputField,
     type PinInputFieldProps,
-    DropdownSelect,
-    type DropdownSelectProps
+    ListBox,
+    type ListBoxProps
 } from "./Form";
 
 // METER ===============================================================================================================